Rename enchant-ispell.c to enchant.c; enchant.c to lib.c
This means that the executable name is the same as the source; no
user-visible change, but smooths a kink out of the build system.
