+++ src/Makefile.am
@@ -14,7 +14,7 @@ else
 libenchant_la_LDFLAGS += -version-info $(VERSION_INFO)
 endif
 
-libenchant_la_SOURCES = enchant.c pwl.c enchant.h pwl.h
+libenchant_la_SOURCES = lib.c pwl.c enchant.h pwl.h
 if OS_WIN32
 libenchant_la_SOURCES += libenchant.rc
 endif
@@ -35,7 +35,6 @@ man_MANS = enchant.1
 
 LDADD = libenchant.la $(ENCHANT_LIBS)
 bin_PROGRAMS = enchant-lsmod enchant
-enchant_SOURCES = enchant-ispell.c
 
 EXTRA_DIST = enchant.i $(ordering_DATA) $(man_MANS)
